Create entity property columns control SN-640
